Reorganize test environment setup
